+++ src/main/java/com/farmstory/service/EmailService.java
@@ -78,9 +78,4 @@ public class EmailService {
             throw new RuntimeException(e);
         }
     }
-
-
-
-
-
 }
