+++ src/main/resources/static/css/user/findPass.css
@@ -1,0 +1,353 @@
+* {
+    margin: 0;
+    padding: 0;
+    font-family: "Inter", sans-serif;
+    box-sizing: border-box;
+    font-size: 12px;
+    font-weight: 400;
+}
+
+input:focus{outline: none;}
+
+#wrapper {
+    width: 100%;
+    height: auto;
+    top: -203px;
+    left: -480px;
+    background: #FFFFFF;
+}
+
+
+header {
+    width: 100%;
+}
+
+header > div {
+    width: 790px;
+    height: 30px;
+    margin: 0 auto;
+}
+
+header > div > .top {
+    font-size: 10.13px;
+    line-height: 30px;
+    text-align: left;
+}
+
+input::placeholder{
+    text-indent: 6px;
+} /* 텍스트 들여쓰기 */
+
+input{
+    background: #E4EAEC;
+}
+
+main {
+    margin-top: 65px;
+    margin-bottom: 65px;
+    position: relative;
+}
+
+main > section {
+    margin: 0 auto;
+    width: 600px;
+    height: 306px;
+    position: relative;
+}
+
+main > section > form {
+    position: absolute;
+    position: relative;
+    margin: 0 auto;
+    width: 600px;
+    height: 161px;
+    top: 16px;
+    padding: 1px 0px 0px 0px;
+}
+
+main > section > form > h1 {
+    display: inline-block;
+    position: relative;
+    position: absolute;
+    top: 1px;
+    width: 75.2px;
+    height: 14px;
+    line-height: 14.52px;
+    text-align: left;
+    color: #000000;
+    white-space: nowrap; /* 한 줄에 텍스트를 표시 */
+    overflow: visible; /* 텍스트가 잘리지 않도록 설정 */
+}
+
+main > section > form > table {
+    position: absolute;
+    position: relative;
+    width: 600px;
+    height: 138px;
+    top: 23px;
+    border: 2px 0px 0px 0px;
+    border-top: 2px solid #111111;
+    border-collapse: collapse;
+}
+
+main > section > form > table > tbody {
+    position: relative;
+    position: absolute;
+    width: 599px;
+    height: 136.5px;
+    top: 1px;
+    left: 0.5px;
+}
+
+.password > tbody > tr:nth-child(1) > th {
+    position: relative;
+    width: 120px;
+    height: 49.5px;
+    padding: 0px 71.3px 0px 12.5px;
+    border: 1px 0px 0px 0px;
+    border: 1px solid #E9E9E9;
+}
+
+.password > tbody > tr:nth-child(1) > th > label {
+    position: absolute;
+    width: 36.2px;
+    height: 14px;
+    top: 17.5px;
+    left: 12.5px;
+    font-size: 12px;
+    font-weight: 400;
+    line-height: 14.52px;
+    text-align: left;
+    color: #000000;
+}
+
+.password > tbody > tr:nth-child(1) > td {
+    width: 479px;
+    height: 49.5px;
+    left: 120px;
+    border: 1px 0px 0px 0px;
+    border: 1px solid #E9E9E9;
+}
+
+.password > tbody > tr:nth-child(1) > td > input {
+    width: 164px;
+    height: 24px;
+    top: 13px;
+    margin-left: 12.5px;
+    padding: 3.5px 1px 3.5px 1px;
+    border: 1px 0px 0px 0px;
+    background: #F7F7F7;
+    border: 1px solid #E4EAEC;
+
+}
+
+.password > tbody > tr:nth-child(2) > th {
+    position: relative;
+    width: 120px;
+    height: 87px;
+    padding: 0px 71.3px 0px 12.5px;
+    border: 1px 0px 0px 0px;
+    border: 1px solid #E9E9E9;
+}
+
+.password > tbody > tr:nth-child(2) > th > label {
+    position: absolute;
+    width: 36.2px;
+    height: 14px;
+    top: 36px;
+    left: 12.5px;
+    font-size: 12px;
+    font-weight: 400;
+    line-height: 14.52px;
+    text-align: left;
+    color: #000000;
+    display: inline-block;
+}
+
+.password > tbody > tr:nth-child(2) > td {
+    width: 479px;
+    height: 87px;
+    border: 1px 0px 0px 0px;
+    border: 1px solid #E9E9E9;
+    position: relative;
+}
+
+.password > tbody > tr:nth-child(2) > td > div {
+    display: block;
+}
+
+
+.password > tbody > tr:nth-child(2) > td > div:nth-of-type(1) > input {
+    position: absolute;
+    width: 164px;
+    height: 24px;
+    top: 18px;
+    left: 12.5px;
+    padding: 3.5px 1px 3.5px 1px;
+    border: 1px 0px 0px 0px;
+    background: #F7F7F7;
+    border: 1px solid #E4EAEC;
+
+}
+
+.password > tbody > tr:nth-child(2) > td .auth-button {
+    position: absolute;
+    width: 85px;
+    height: 27px;
+    top: 16.5px;
+    left: 179.5px;
+    border: 1px 0px 0px 0px;
+    font-size: 12px;
+    font-weight: 400;
+    background: #EFEFEF;
+    border: 1px solid #BEBEBE;
+
+}
+
+.password > tbody > tr:nth-child(2) > td .auth-button > span {
+    text-align: center;
+    width: 75.2px;
+    height: 14px;
+    top: 6px;
+    left: 5px;
+    line-height: 14.52px;
+
+}
+
+.password > tbody > tr:nth-child(2) > td > div:nth-of-type(2) > input {
+    position: absolute;
+    width: 164px;
+    height: 24px;
+    top: 49px;
+    left: 12.5px;
+    padding: 3.5px 1px 3.5px 1px;
+    gap: 0px;
+    border: 1px 0px 0px 0px;
+    opacity: 0px;
+    background: #F7F7F7;
+    border: 1px solid #E4EAEC;
+
+}
+
+.password > tbody > tr:nth-child(2) > td .ok-button {
+    position: absolute;
+    width: 34px;
+    height: 27px;
+    top: 47.5px;
+    left: 179.5px;
+    gap: 0px;
+    border: 1px 0px 0px 0px;
+    opacity: 0px;
+    background: #EFEFEF;
+    border: 1px solid #BEBEBE;
+    font-family: Inter;
+    font-size: 12px;
+    font-weight: 400;
+    line-height: 14.52px;
+    text-align: center;
+    color: #000000;
+}
+
+.password > tbody > tr:nth-child(2) > td .ok-button > span {
+    text-align: center;
+    width: 24.2px;
+    height: 14px;
+    top: 6px;
+    left: 5px;
+}
+
+main > section > form > p {
+    position: absolute;
+    width: 440px;
+    height: 60px;
+    top: 173px;
+    gap: 0px;
+    opacity: 0px;
+    font-size: 12px;
+    font-weight: 400;
+    line-height: 20px;
+    text-align: left;
+    color: #000000;
+}
+
+.cancel-button {
+    position: absolute;
+    width: 42px;
+    height: 35px;
+    top: 255px;
+    left: 513px;
+    gap: 0px;
+    border: 1px 0px 0px 0px;
+    opacity: 0px;
+    border: 1px solid #BEBEBE;
+    font-size: 12px;
+    font-weight: 400;
+    line-height: 14.52px;
+    text-align: center;
+    background: white;
+}
+
+.cancel-button > span {
+    width: 42px;
+    height: 35px;
+    top: 255px;
+    left: 513px;
+    gap: 0px;
+    border: 1px 0px 0px 0px;
+    opacity: 0px;
+    line-height: 14.52px;
+    text-align: center;
+    background: white;
+}
+
+.next-button {
+    position: absolute;
+    width: 42px;
+    height: 35px;
+    top: 255px;
+    left: 558px;
+    gap: 0px;
+    border: 1px 0px 0px 0px;
+    opacity: 0px;
+    border: 1px solid #BEBEBE;
+    font-size: 12px;
+    font-weight: 400;
+    line-height: 14.52px;
+    text-align: center;
+    background: white;
+}
+
+.next-button > span {
+    width: 42px;
+    height: 35px;
+    top: 255px;
+    left: 513px;
+    gap: 0px;
+    border: 1px 0px 0px 0px;
+    opacity: 0px;
+    line-height: 14.52px;
+    text-align: center;
+    background: white;
+}
+
+
+footer {
+    height: 17px;
+    width: 100%;
+    margin: 0 auto;
+    margin-top: 60px;
+    color: #777777;
+}
+
+footer > div {
+    width: 790px;
+    margin: 0 auto;
+}
+
+footer > div > span:nth-of-type(1) {
+    float: left;
+}
+
+footer > div > span:nth-of-type(2) {
+    float: right;
+}
