+++ src/main/resources/templates/user/findPassChange.html
@@ -9,42 +9,29 @@
 </head>
 <body>
 <div id="wrapper">
-    <!-- 헤더 -->
-    <header>
-        <div>
-            <span class="top">Board System</span>
-        </div>
-    </header>
+    <th:block th:include ="/include/header.html"/>
     <!-- 메인 -->
     <main>
         <section>
-            <form action="">
+            <form th:action="@{/user/findPassChange}" method="post" id="passwordResetForm">
                 <h1>비밀번호 변경</h1>
                 <table class="passwordchange">
                     <tbody>
                     <tr>
-                        <th>
-                            <label>아이디</label>
-                        </th>
-                        <td>
-                            honggildong
-                        </td>
-                    </tr>
                     <tr>
-                        <th>
-                            <label>새 비밀번호</label>
-                        </th>
-                        <td>
-                            <input type="text" placeholder="새 비밀번호 입력">
-                        </td>
-                    </tr>
-                    <tr>
-                        <th>
-                            <label>새 비밀번호 확인</label>
-                        </th>
-                        <td>
-                            <input type="text" placeholder="새 비밀번호 입력">
+                        <th>아이디</th>
+                        <input class="uid" name="uid" th:value="${uid}"></input>
 
+                    </tr>
+                    <tr>
+                        <th>새 비밀번호</th>
+                        <td><input type="password" id="newPassword" name="pass" placeholder="새 비밀번호 입력"></td>
+                    </tr>
+                    <tr>
+                        <th>새 비밀번호 확인</th>
+                        <td>
+                            <input type="password" id="confirmPassword" name="confirmPassword" placeholder="새 비밀번호 확인">
+                            <span id="resultPass"  class="validationMessage"></span>
                         </td>
                     </tr>
                     </tbody>
@@ -53,30 +40,64 @@
                     비밀번호를 변경해 주세요.<br>
                     영문, 숫자, 특수문자를 사용하여 8자 이상 입력해 주세요.
                 </p>
-                <div>
-                    <button type="button" class="cancel-button">
-                        <span>취소</span>
-                    </button>
-                    <button type="button" class="next-button">
-                        <span>다음</span>
-                    </button>
+                <div class="btn-group">
+                    <button type="button" class="btnCancel cancel-button" onclick="history.back();"><span>취소</span></button>
+                    <button type="submit" class="btnNext next-button"><span>다음</span></button>
                 </div>
             </form>
         </section>
-    </main>
+    </main><!--main end-->
 
-    <!-- 푸터 -->
-    <footer id="footer">
-        <div>
-                <span>
-                    Copyright© 김철학(개발에반하다.)
-                </span>
-            <span>
-                    v1.0.1
-                </span>
-        </div>
-    </footer>
-</div>
+    <th:block th:insert ="~{/include/footer.html}"/>
+</div>!--wrapper end-->
 
 </body>
 </html>
+
+
+<script>
+    // 새 비밀번호 유효성 검사에 사용할 정규표현식
+    const rePass = /^(?=.*[a-zA-z])(?=.*[0-9])(?=.*[$`~!@$!%*#^?&\\(\\)\-_=+]).{5,16}$/;
+
+    // 새 비밀번호 유효성 검사에 사용할 상태변수
+    let isPassOk = false;
+
+    window.onload=function (){
+
+        const registerForm = document.getElementsByTagName('form')[0];
+        const resultPass = document.getElementsByClassName('resultPass')[0];
+        const submit = document.querySelector('.next-button');
+
+        // 2.비밀번호 유효성 검사
+        registerForm.pass2.addEventListener('focusout', function(){
+
+            const pass = registerForm.userPass.value;
+            const pass2 = registerForm.pass2.value;
+
+            if(!pass.match(rePass)){
+                resultPass.innerText = "비밀번호가 유효하지 않습니다.";
+                resultPass.style.color = 'red';
+                return;
+            }
+
+            if(pass == pass2){
+                resultPass.innerText = "비밀번호가 일치합니다.";
+                resultPass.style.color = 'green';
+                isPassOk = true;
+            }else{
+                resultPass.innerText = "비밀번호가 일치하지 않습니다.";
+                resultPass.style.color = 'red';
+
+            }
+        });
+
+        submit.addEventListener('click', function (e) {
+
+            if (!isPassOk) {
+                alert('비밀번호가 유효하지 않습니다. 다시 입력해 주세요.');
+                e.preventDefault(); // 제출 중단
+            }
+        });
+
+    }
+</script>
