+++ src/main/resources/templates/user/register.html
@@ -21,7 +21,8 @@
         <main>
             <div class="mainIn">
                 <section class="register">
-                    <form th:action="@{/user/register}" method="post">
+
+                    <form th:action="@{/user/register}" name="registerForm" method="post">
                         <table border="1">
                             <caption>사이트 이용정보 입력</caption>
                             <tr>
